# backend/openapi.yaml
openapi: 3.0.3
info:
  title: IDSC Finance Management System API
  description: REST API for the IDSC School Finance Management System
  version: 1.0.0
servers:
  - url: http://localhost:5000
...
paths:
...
  /api/v1/receipts:
    get:
      tags:
        - Receipts
      summary: Get all receipts
      responses:
        "200":
          description: List of receipts

    post:
      tags:
        - Receipts
      summary: Create a receipt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiptInput"
      responses:
        "201":
          description: Receipt created successfully
        "400":
          description: Invalid receipt
        "409":
          description: Receipt number already exists
  /api/v1/receipts/{id}:
    get:
      tags:
        - Receipts
      summary: Get one receipt
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Receipt found
        "404":
          description: Receipt not found

    put:
      tags:
        - Receipts
      summary: Update a receipt
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiptInput"
      responses:
        "200":
          description: Receipt updated successfully
        "404":
          description: Receipt not found

    delete:
      tags:
        - Receipts
      summary: Delete a receipt
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Receipt deleted successfully
        "404":
          description: Receipt not found
  /api/v1/receipts/student/{studentId}:
    get:
      tags:
        - Receipts
      summary: Get receipts by student ID
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student receipts
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: MongoDB document ID
      schema:
        type: string
  schemas:
...
    ReceiptInput:
      type: object
      required:
        - receiptNumber
        - paymentId
        - studentId
        - studentName
        - amountPaid
        - paymentMethod
        - previousBalance
        - remainingBalance
        - processedBy
      properties:
        receiptNumber:
          type: string
        paymentId:
          type: string
        studentId:
          type: string
        studentName:
          type: string
        paymentDate:
          type: string
          format: date-time
        amountPaid:
          type: number
        paymentMethod:
          type: string
          enum:
            - Cash
            - Bank Transfer
            - Online Payment
            - Other
        referenceNumber:
          type: string
        previousBalance:
          type: number
        remainingBalance:
          type: number
        processedBy:
          type: string